user-manager/requirements.txt
flask
grpcio
grpcio-tools
PyMySQL
bcrypt
cryptography
prometheus_client
